==> 04_GEN__rome_with_naf_csv.ipynb <==
# BEST NAF match

import pandas as pd
from sentence_transformers import SentenceTransformer, util

# Charger les fichiers CSV
rome_df = pd.read_csv("rome.csv")
naf_df = pd.read_csv("naf_codes_001_desc.csv")

# Nettoyer les colonnes desc (si besoin)
rome_df['desc'] = rome_df['desc'].astype(str)
naf_df['desc'] = naf_df['desc'].astype(str)

# Charger le modèle d'embeddings
model = SentenceTransformer('all-MiniLM-L6-v2')  # petit et efficace pour la similarité
# model = SentenceTransformer("thenlper/gte-large")

# Calculer les embeddings pour NAF
naf_embeddings = model.encode(naf_df['desc'].tolist(), convert_to_tensor=True,  normalize_embeddings=True, show_progress_bar=True)

# Fonction pour trouver le code NAF le plus proche
def find_best_naf(desc):
    desc_embedding = model.encode(desc, convert_to_tensor=True)
    # Calculer la similarité cosine
    similarities = util.cos_sim(desc_embedding, naf_embeddings)
    best_idx = similarities.argmax().item()  # <-- convertir en int
    return naf_df.iloc[best_idx]['code_naf']

# Appliquer à chaque ligne du CSV rome
rome_df['code_naf'] = rome_df['desc'].apply(find_best_naf)

# Sauvegarder le nouveau CSV
# f = "rome_with_naf.csv"
# f = "rome_with_naf__thenlper_gte-large.csv"
f = "rome_with_naf_all-MiniLM-L6-v2.csv"
rome_df.to_csv(f, index=False)
print(f"Fichier '{f}' créé avec succès !")
