==> 05_BUILD__fusion_naf_rome_001_csv.ipynb <==
import polars as pl

# romes = pl.read_csv("rome_with_naf__thenlper_gte-large.csv")
romes = pl.read_csv("rome_with_naf_all-MiniLM-L6-v2.csv")
nafs = pl.read_csv("naf_codes_001_desc.csv")

romes = romes.with_columns([
    pl.col(col).cast(pl.Utf8) for col in romes.columns
])

nafs = nafs.with_columns([
    pl.col(col).cast(pl.Utf8) for col in nafs.columns
])

print("%s   %s" % (type(romes), type(nafs)))
print(romes.columns)
print(nafs.columns)
fusion_df = pl.concat([romes, nafs])
print("Nombre d'éléments: ", len(fusion_df))
# fusion_df.write_csv("fusion_naf_rome_001_thenlper_gte-large.csv")
fusion_df.write_csv("fusion_naf_rome_001_allMiniLM_L6_v2.csv")
fusion_df.sample(20)
